Minor refactoring around :data
- Fix path: src/main/java/dk/nodes/template/data
- Move /storage/PrefManagerImpl from :app to :data
