+++ app/src/main/java/dk/nodes/template/injection/modules/RestRepositoryBinding.kt
@@ -2,7 +2,7 @@ package dk.nodes.template.injection.modules
 
 import dagger.Binds
 import dagger.Module
-import dk.nodes.template.network.RestPostRepository
+import dk.nodes.template.data.network.RestPostRepository
 import dk.nodes.template.domain.repositories.PostRepository
 import javax.inject.Singleton
