+++ data/src/main/java/dk/nodes/template/data/storage/PrefManagerImpl.kt
@@ -1,9 +1,8 @@
-package dk.nodes.template.storage
+package dk.nodes.template.data.storage
 
 import android.content.Context
 import android.content.SharedPreferences
 import androidx.core.content.edit
-import androidx.preference.PreferenceManager
 import dk.nodes.template.domain.managers.PrefManager
 import javax.inject.Inject
